lib/chord-metadata/chord_metadata_service/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
Babel==2.9.0
bento-lib==0.11.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
codecov==2.1.11
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.4.6
distlib==0.3.1
Django==2.2.17
django-cors-headers==3.7.0
django-filter==2.4.0
django-nose==1.4.7
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
docutils==0.16
elasticsearch==7.8.0
entrypoints==0.3
fhirclient==3.2.0
filelock==3.0.12
flake8==3.8.4
idna==2.10
imagesize==1.2.0
importlib-metadata==3.5.0
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.7.0
nose==1.3.7
openapi-codec==1.3.2
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT[crypto]==1.7.1
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
PyYAML==5.4.1
rdflib==5.0.0
rdflib-jsonld==0.5.0
redis==3.5.3
requests==2.25.1
rfc3987==1.3.8
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
strict-rfc3339==0.7
toml==0.10.2
tox==3.22.0
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
